split group list page out as separate component
