+++ src/libs/routes.js
@@ -6,13 +6,13 @@ import * as AppLauncher from 'src/analysis/AppLauncher';
 import * as SignOutPage from 'src/auth/signout/SignOutPage';
 import * as DataBrowserDetails from 'src/data-catalog/DataBrowserDetails';
 import * as DataBrowserPreview from 'src/data-catalog/DataBrowserPreview';
+import * as Group from 'src/groups/Group';
 import * as ImportDataPage from 'src/import-data/ImportDataPage';
 import { routeHandlersStore } from 'src/libs/state';
 import * as Projects from 'src/pages/billing/BillingListPage';
 import * as Environments from 'src/pages/EnvironmentsPage/EnvironmentsPage';
 import * as FeaturePreviews from 'src/pages/FeaturePreviews';
-import * as Group from 'src/pages/groups/Group';
-import * as Groups from 'src/pages/groups/List';
+import * as Groups from 'src/pages/groups/GroupListPage';
 import * as ImportWorkflow from 'src/pages/ImportWorkflow/ImportWorkflow';
 import * as LandingPage from 'src/pages/LandingPage';
 import * as Code from 'src/pages/library/Code';
